Modify Qt5 drivers so they work with PythonQt.
From https://sourceforge.net/p/pythonqt/discussion/631393/thread/7b5e4b43/

MeVisLab 2.3 contains a matplotlib backend that was ported to PythonQt, by coincidence I did that port. It is not implemented as a full backend of its own, just a patch that replaces calls to PyQt/PySide with PythonQt in the Qt4 backend.

I did not have time to clean it up and to provide it back to the matplotlib community. It is fully functional except for the figure options dialog.

You can get it by downloading MeVisLab 2.3 RC and going to

INSTALLDIR/Packages/MeVis/ThirdParty/Python/…/Libs/matplotlib/…

(or Libs/site-packages if you are downloading 2.3 Stable Release)

and taking the matplotlib/backends/ directory. You can diff that against your matplotlib version to see the changes.

Florian Link - 2012-07-18